agents/FiMoneyAgent/pyproject.toml
[project]
name = "fi-money-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.16",
    "click>=8.2.1",
    "google-adk>=1.8.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
